tests: organize existing tests into new categories
