+++ tests/conftest.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+from tests.fixtures.collector import *
+from tests.fixtures.writer import *
